Update LoRa Basic Example
Changed the int before prevWindTelemetry to an unsigned long.

+++ examples/iotQi-LoRa-basic/iotQi-LoRa-basic.ino
@@ -43,7 +43,7 @@ void setup() {
 
 
 /* loop() ------------------------------------------------------------------------------------------------------------- */
-int prevWindTelemetry = 0;
+unsigned long prevWindTelemetry = 0;
 
 void loop()
 {
